requirements.txt
ollama

mysql-connector-python

transformers

pandas

torch

streamlit
